engine/requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
bitsandbytes==0.39.1
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
datasets==2.13.1
dill==0.3.6
einops==0.6.1
et-xmlfile==1.1.0
filelock==3.12.2
fire==0.5.0
frozenlist==1.3.3
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
lit==16.0.6
loralib==0.1.1
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
peft==0.4.0
protobuf==3.19.0
psutil==5.9.5
pyarrow==12.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scipy==1.11.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
termcolor==2.3.0
tokenizers==0.14.0
torch==2.0.1
tqdm==4.65.0
transformers==4.34.0
transformers-stream-generator==0.0.4
triton==2.0.0
typing_extensions==4.7
tzdata==2023.3
urllib3==2.0.3
xxhash==3.2.0
yarl==1.9.2
pytest==7.4.1
pytest-mock==3.11.1
evaluate==0.4.0
responses==0.18.0
rouge-score==0.1.2
contourpy==1.1.1
joblib==1.3.2
pyparsing==3.1.1
matplotlib==3.8.0
pluggy==1.3.0
absl-py==1.4.0
click==8.1.7
cycler==0.11.0
iniconfig==2.0.0
kiwisolver==1.4.5
exceptiongroup==1.1.3
rouge==1.0.1
nltk==3.8.1
fonttools==4.42.1
bleu==0.3
Pillow==10.0.1
efficiency==1.31
tomli==2.0.1
langchain-core==0.1.18
langchain==0.1.5
openai==1.11.1
mininet==2.3.0.dev6
